Actualizacion_owners

+++ MyVet.Web/Controllers/PetTypesController.cs
@@ -124,25 +124,26 @@ namespace MyVet.Web.Controllers
             }
 
             var petType = await _context.PetTypes
-                .FirstOrDefaultAsync(m => m.PetTypeId == id);
+                .Include(pt=>pt.Pets)               
+                .FirstOrDefaultAsync(pt => pt.PetTypeId == id);
             if (petType == null)
             {
                 return NotFound();
             }
 
-            return View(petType);
-        }
-
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
+            if (petType.Pets.Count > 0)
+            {
+                ModelState.AddModelError(string.Empty, "The Pet type can´t be removed");
+                return RedirectToAction(nameof(Index));
+            }
+           
             _context.PetTypes.Remove(petType);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
+
         }
+
+
 
         private bool PetTypeExists(int id)
         {
